Update README.md
In this exercise, I loaded the sakila datasets and tried on some of the exercices with this dataset. The exercise includes JOIN, WHERE, INSERT INTO, VIEWS, etc.
It will be my thoughts and my own code for sakila exercise, if you have better solutions or one liner, really hope to share and improve with you guys.